Refactor(sendRequests)!: sendRequest is now fully dependant of python3
